backend/requirements.txt
# ── All Python backend services share these dependencies ─────────────────────

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database (async SQLAlchemy + PostgreSQL driver)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0

# Read .env files
python-dotenv==1.0.1

# gRPC (Inventory gRPC server + clients in Claim/Payment services)
grpcio==1.67.1
grpcio-tools==1.67.1

# RabbitMQ (Claim/Notification/Auditor event publishing — blocking)
pika==1.3.2
# RabbitMQ async client used by Claim/Payment orchestrators
aio-pika==9.4.3

# Geohash encoding/decoding for location-based search
geohash2==1.1

# JWT decode/encode (OutSystems wrapper + shared jwt_auth)
PyJWT==2.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
aiosqlite==0.20.0

# JWT verification (RS256) — used by shared/jwt_auth.py
PyJWT==2.9.0
cryptography==43.0.3